2 States, Application Control, Display Generations
Added a 2nd state to denote new cells vs old cells

Application state now controls most of the login instead of just the
Board. Did this so we can have other components tell the application
what is going on and then it informs the Board.

The Generation number is now displayed